# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Palace HTTP API",
    "description": "Owner-scoped Palace API. Schema describes wire structure, not authorization or transactional guarantees. Body limits use injected ImportLimits (default raw history 8 MiB); HTTP limit is bytes * 6 + 64 KiB. UTF-8 byte limits, nonblank strings and canonical cursor upper bounds are enforced by the server. Native extractor errors may be text/plain. No deployment credentials or environment metadata are included.",
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
    "/api/conversations/{id}": {
      "get": {
        "tags": [
          "business"
        ],
        "summary": "Returns a stable tree and controlled external link only for the authenticated owner.",
        "operationId": "conversation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Owner-scoped UUID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "no-store"
              },
              "Set-Cookie": {
                "schema": {
                  "type": "string"
                },
                "description": "Refreshed production session; absent in fixed-user mode"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationDetail"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input; JSON domain error or native extractor text as declared",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InputErrorResponse"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource unavailable in owner scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Identity, source session or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal persistence or response failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Identity provider temporarily unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "delete": {
        "tags": [
          "path_management"
        ],
        "summary": "Deletes the card and all branches atomically inside the authenticated owner's scope.",
        "operationId": "delete_conversation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Owner-scoped UUID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Origin",
            "in": "header",
            "description": "Exactly one header matching the configured external origin; null, duplicates and omissions are rejected",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "no-store"
              },
              "Set-Cookie": {
                "schema": {
                  "type": "string"
                },
                "description": "Refreshed production session; absent in fixed-user mode"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Deleted"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input; JSON domain error or native extractor text as declared",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Origin rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource unavailable in owner scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Identity, source session or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal persistence or response failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Identity provider temporarily unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
    "/api/import": {
      "post": {
        "tags": [
          "business"
        ],
        "summary": "Adapts pasted JSON to the shared byte parser; owner fields are not part of this protocol.",
        "operationId": "import_text",
        "parameters": [
          {
            "name": "Origin",
            "in": "header",
            "description": "Exactly one header matching the configured external origin; null, duplicates and omissions are rejected",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextImport"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "no-store"
              },
              "Set-Cookie": {
                "schema": {
                  "type": "string"
                },
                "description": "Refreshed production session; absent in fixed-user mode"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResult"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input; JSON domain error or native extractor text as declared",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InputErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Origin rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource unavailable in owner scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Identity, source session or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Input capacity exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InputErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal persistence or response failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Identity provider temporarily unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/import/file": {
      "post": {
        "tags": [
          "business"
        ],
        "summary": "Accumulates a bounded file body and sends precisely the same bytes through the common parser.",
        "operationId": "import_file",
        "parameters": [
          {
            "name": "Origin",
            "in": "header",
            "description": "Exactly one header matching the configured external origin; null, duplicates and omissions are rejected",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/FileImport"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "no-store"
              },
              "Set-Cookie": {
                "schema": {
                  "type": "string"
                },
                "description": "Refreshed production session; absent in fixed-user mode"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResult"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input; JSON domain error or native extractor text as declared",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InputErrorResponse"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Origin rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource unavailable in owner scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Identity, source session or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "Input capacity exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InputErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal persistence or response failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Identity provider temporarily unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Conversation": {
        "type": "object",
        "required": [
          "id",
          "owner_id",
          "source"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "owner_id": {
            "type": "string",
            "format": "uuid"
          },
          "source": {
            "$ref": "#/components/schemas/Source"
          }
        }
      },
      "ConversationDetail": {
        "type": "object",
        "required": [
          "conversation",
          "messages",
          "paths"
        ],
        "properties": {
          "conversation": {
            "$ref": "#/components/schemas/Conversation"
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Message"
            }
          },
          "paths": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConversationPath"
            }
          }
        }
      },
...
      "Deleted": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "ErrorCode": {
        "type": "string",
        "enum": [
          "authentication_required",
          "origin_rejected",
          "identity_or_request_conflict",
          "session_already_exists",
          "not_found",
          "identity_unavailable",
          "persistence_failed"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error",
          "login"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "login": {
            "type": "string",
            "description": "/auth/login for authentication_required; the empty string for other errors. Always present."
          }
        }
      },
...
      "ImportResult": {
        "type": "object",
        "required": [
          "import_id",
          "conversation_id",
          "path_id",
          "head_message_id",
          "created",
          "reused"
        ],
        "properties": {
          "conversation_id": {
            "type": "string",
            "format": "uuid"
          },
          "created": {
            "type": "integer",
            "minimum": 0
          },
          "head_message_id": {
            "type": "string",
            "format": "uuid"
          },
          "import_id": {
            "type": "string",
            "format": "uuid"
          },
          "path_id": {
            "type": "string",
            "format": "uuid"
          },
          "reused": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "InputErrorKind": {
        "type": "string",
        "enum": [
          "syntax",
          "field",
          "limit"
        ]
      },
      "InputErrorResponse": {
        "type": "object",
        "required": [
          "kind",
          "path",
          "message"
        ],
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/InputErrorKind"
          },
          "message": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        }
      },
...
      "Source": {
        "type": "string",
        "enum": [
          "chatgpt",
          "gemini",
          "grok"
        ]
      },
...
      "TextImport": {
        "type": "object",
        "description": "Original history JSON is a string here, not an embedded message array.",
        "required": [
          "title",
          "occurred_at",
          "source",
          "session_id",
          "history",
          "idempotency_key"
        ],
        "properties": {
          "history": {
            "type": "string",
            "description": "Nonempty role/content array encoded as text. Default limits: 8 MiB, 10000 messages,\n1 MiB per content, depth 32. Extra message fields are ignored; content is preserved."
          },
          "idempotency_key": {
            "type": "string",
            "description": "Nonblank after trimming; at most 128 UTF-8 bytes, scoped to the owner."
          },
          "occurred_at": {
            "type": "integer",
            "format": "int64",
            "description": "User-selected Unix epoch milliseconds, calendar years 0001 through 9999.",
            "maximum": 253402300799999,
            "minimum": -62135596800000
          },
          "session_id": {
            "type": "string",
            "description": "1..512 UTF-8 bytes; one source path segment, no whitespace, controls or /\\\\?#%:."
          },
          "source": {
            "$ref": "#/components/schemas/Source"
          },
          "title": {
            "type": "string",
            "description": "Nonblank after trimming; preserve the original value; at most 1024 UTF-8 bytes."
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "login_flow": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-palace-login",
        "description": "Browser-bound one-time login flow; validated together with query state before exchanging code."
      },
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-palace-session",
        "description": "Production browser session; Secure, HttpOnly, SameSite=Lax, Path=/. Business handlers receive a server-resolved owner. Local fixed-user mode is a development-only boundary."
      }
    }
  }
}